update request_controller_typed_link and tests
