+++ ferry/lib/src/request_controller_typed_link.dart
@@ -3,7 +3,6 @@ import 'package:rxdart/rxdart.dart';
 import 'package:ferry_exec/ferry_exec.dart';
 
 class RequestControllerTypedLink extends TypedLink {
-  /// A stream controller that handles all [OperationRequest]s.
   final StreamController<OperationRequest> requestController;
 
   RequestControllerTypedLink([
@@ -16,33 +15,47 @@ class RequestControllerTypedLink extends TypedLink {
     forward,
   ]) {
     var initial = true;
-    var prev = BehaviorSubject<OperationResponse<TData, TVars>>.seeded(null);
+    ValueConnectableStream<OperationResponse<TData, TVars>> prev;
+    final subs = <StreamSubscription>[];
+
+    final cancelSubs = () => subs
+      ..forEach((sub) => sub.cancel())
+      ..clear();
 
     return requestController.stream
         .whereType<OperationRequest<TData, TVars>>()
         .where((req) => request.requestId == req.requestId)
+        .doOnDone(cancelSubs)
         .switchMap(
-          (req) => prev = req.updateResult == null
-              ? forward(req).shareValue()
-              : CombineLatestStream.combine2<
-                  OperationResponse<TData, TVars>,
-                  OperationResponse<TData, TVars>,
-                  OperationResponse<TData, TVars>>(
-                  prev,
-                  forward(req),
-                  (previous, response) => OperationResponse(
-                    operationRequest: response.operationRequest,
-                    data: response.operationRequest.updateResult(
-                      previous?.data,
-                      response.data,
-                    ),
-                    dataSource: response.dataSource,
-                    linkException: response.linkException,
-                    graphqlErrors: response.graphqlErrors,
-                  ),
-                ).shareValue(),
-        )
-        .doOnListen(
+      (req) {
+        StreamSubscription sub;
+        Stream<OperationResponse<TData, TVars>> stream;
+        if (req.updateResult == null) {
+          stream = forward(req);
+          cancelSubs();
+        } else {
+          stream = CombineLatestStream.combine2<OperationResponse<TData, TVars>,
+              OperationResponse<TData, TVars>, OperationResponse<TData, TVars>>(
+            prev ?? Stream.value(null),
+            forward(req),
+            (previous, response) => OperationResponse(
+              operationRequest: response.operationRequest,
+              data: response.operationRequest.updateResult(
+                previous?.data,
+                response.data,
+              ),
+              dataSource: response.dataSource,
+              linkException: response.linkException,
+              graphqlErrors: response.graphqlErrors,
+            ),
+          );
+        }
+        prev = stream.doOnDone(() => sub.cancel()).publishValue();
+        sub = prev.connect();
+        subs.add(sub);
+        return prev;
+      },
+    ).doOnListen(
       () {
         if (initial && request.executeOnListen) {
           scheduleMicrotask(() => requestController.add(request));
