+++ ferry/test/link_subscription_cancel_test.dart
@@ -1,11 +1,10 @@
 import 'dart:async';
 
+import 'package:ferry/typed_links.dart';
 import 'package:gql_link/gql_link.dart';
 import 'package:gql_exec/gql_exec.dart';
 import 'package:gql_exec/src/response.dart';
 import 'package:test/test.dart';
-
-import 'package:ferry/ferry.dart';
 
 import 'package:ferry_test_graphql/queries/variables/human_with_args.req.gql.dart';
 
@@ -42,7 +41,7 @@ void main() {
       'Stream in Link is cancelled when no one is listening - for all FetchPolicies',
       () async {
     final link = _StreamCancelTestLink();
-    final client = Client(link: link);
+    final client = FetchPolicyTypedLink(link: link);
     // test all fetchpolicies except CacheOnly - CacheOnly does not actually the link
     for (final fetchPolicy in [
       FetchPolicy.NetworkOnly,
